Updated ALL to latest 1.4.350.0 version. Removed oldest and useless version 1.2.198.0. Fixed some linter warnings and other stuff
