+++ recipes/spirv-cross/config.yml
@@ -1,5 +1,5 @@
 versions:
-  "1.4.341.0":
+  "1.4.350.0":
     folder: all
   "1.4.321.0":
     folder: all
@@ -20,6 +20,4 @@ versions:
   "1.3.243.0":
     folder: all
   "1.3.239.0":
-    folder: all
-  "cci.20211113":
     folder: all
