+++ recipes/spirv-headers/config.yml
@@ -1,5 +1,5 @@
 versions:
-  "1.4.341.0":
+  "1.4.350.0":
     folder: all
   "1.4.313.0":
     folder: all
@@ -14,6 +14,4 @@ versions:
   "1.3.243.0":
     folder: all
   "1.3.239.0":
-    folder: all
-  "1.2.198.0":
     folder: all
