+++ recipes/spirv-reflect/config.yml
@@ -1,5 +1,5 @@
 versions:
-  "1.4.341.0":
+  "1.4.350.0":
     folder: all
   "1.4.313.0":
     folder: all
