+++ recipes/vk-bootstrap/all/conandata.yml
@@ -1,10 +1,10 @@
 sources:
-  "1.4.341":
-    url: "https://github.com/charles-lunarg/vk-bootstrap/archive/refs/tags/v1.4.341.tar.gz"
-    sha256: "c6fcf8befc62bcf7ec4d847cec093522bb470dad15f07218c3102ed892777dc5"
+  "1.4.350":
+    url: "https://github.com/charles-lunarg/vk-bootstrap/archive/refs/tags/v1.4.350.tar.gz"
+    sha256: "848637567c472ee334ea7b694e82c5c9f8966c9a8fa8e687a985223cae2c0c1e"
   "1.3.296":
     url: "https://github.com/charles-lunarg/vk-bootstrap/archive/refs/tags/v1.3.296.tar.gz"
     sha256: "fbff2746134459648a488588a71609d11e6e00f591ac4a3cc6683a131bf31a53"
 vulkan_headers_dependency:
-  "1.4.341": "1.4.341.0"
+  "1.4.350": "1.4.350.0"
   "1.3.296": "1.3.296.0"
