+++ recipes/vk-bootstrap/config.yml
@@ -1,5 +1,5 @@
 versions:
-  "1.4.341":
+  "1.4.350":
     folder: all
   "1.3.296":
     folder: all
