+++ recipes/volk/all/conandata.yml
@@ -1,7 +1,7 @@
 sources:
-  "1.4.341.0":
-    url: "https://github.com/zeux/volk/archive/refs/tags/vulkan-sdk-1.4.341.0.tar.gz"
-    sha256: "42df539c70ffdaea259e317aef73524512f4093f6f4dafb36fa6cf2680c823b9"
+  "1.4.350.0":
+    url: "https://github.com/zeux/volk/archive/refs/tags/vulkan-sdk-1.4.350.0.tar.gz"
+    sha256: "a04f26f76e9a4f9acf936bd2c159f5c4c8348f8ebaf118ff72ba6a9637ad3e80"
   "1.4.313.0":
     url: "https://github.com/zeux/volk/archive/refs/tags/vulkan-sdk-1.4.313.0.tar.gz"
     sha256: "d86bcf1aff499f41a3e445b55df5e393a5ce49b1bda689eb7335b0a0a54a3c0b"
@@ -23,6 +23,3 @@ sources:
   "1.3.239.0":
     url: "https://github.com/zeux/volk/archive/refs/tags/sdk-1.3.239.0.tar.gz"
     sha256: "b628ae7d4e554c20763ac34e1fd18b87725c0c346e2a7106fa23701877bbac65"
-  "1.2.198":
-    url: "https://github.com/zeux/volk/archive/refs/tags/1.2.198.tar.gz"
-    sha256: "8f3869d94ccce265eaef064b1be0af2d8de925e512d3413c1cf02591b99597e2"
