+++ recipes/vulkan-utility-libraries/all/conandata.yml
@@ -1,4 +1,4 @@
 sources:
-  "1.4.341.0":
-    url: "https://github.com/KhronosGroup/Vulkan-Utility-Libraries/archive/refs/tags/vulkan-sdk-1.4.341.0.tar.gz"
-    sha256: "4438cd451b51b5cd13de924bd9d5015c35a06a69e4423452edf79bad646f0469"
+  "1.4.350.0":
+    url: "https://github.com/KhronosGroup/Vulkan-Utility-Libraries/archive/refs/tags/vulkan-sdk-1.4.350.0.tar.gz"
+    sha256: "19a215e9469df0749d7c1b389fc667a3f7e160f0b6da71000fadc30140494563"
