+++ recipes/vulkan-utility-libraries/all/conanfile.py
@@ -1,5 +1,6 @@
 from conan import ConanFile
-from conan.tools.files import copy, get
+from conan.tools.build import check_min_cppstd
+from conan.tools.files import copy, get, rmdir, replace_in_file
 from conan.tools.cmake import cmake_layout, CMake, CMakeToolchain, CMakeDeps
 import os
 
@@ -32,14 +33,22 @@ class VulkanUtilityLibrariesConan(ConanFile):
         self.requires(f"vulkan-headers/{self.version}", transitive_headers=True)
 
     def build_requirements(self):
-        self.tool_requires("cmake/[>=3.22]")
+        self.tool_requires("cmake/[>=3.22.1]")
+
+    def validate(self):
+        check_min_cppstd(self, 17)
 
     def source(self):
         get(self, **self.conan_data["sources"][self.version], strip_root=True)
+        for text in ["set(CMAKE_CXX_STANDARD 17)", "set(CMAKE_CXX_STANDARD_REQUIRED ON)",
+                     "set(CMAKE_POSITION_INDEPENDENT_CODE ON)"]:
+            replace_in_file(self, os.path.join(self.source_folder, "CMakeLists.txt"),
+                            text, "")
 
     def generate(self):
         tc = CMakeToolchain(self)
-        tc.cache_variables["PACKAGE_BUILD_TESTS"] = False
+        tc.cache_variables["BUILD_TESTS"] = False
+        tc.cache_variables["VUL_ENABLE_INSTALL"] = True
         tc.generate()
 
         deps = CMakeDeps(self)
@@ -54,6 +63,7 @@ class VulkanUtilityLibrariesConan(ConanFile):
         copy(self, "LICENSE*", src=self.source_folder, dst=os.path.join(self.package_folder, "licenses"))
         cmake = CMake(self)
         cmake.install()
+        rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
 
     def package_info(self):
         self.cpp_info.set_property("cmake_file_name", "VulkanUtilityLibraries")
