+++ recipes/vulkan-utility-libraries/config.yml
@@ -1,3 +1,3 @@
 versions:
-  "1.4.341.0":
+  "1.4.350.0":
     folder: all
