+++ recipes/vulkan-validationlayers/1.3.x/conandata.yml
@@ -1,0 +1,54 @@
+sources:
+  "1.3.243.0":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.243.0.tar.gz"
+    sha256: "fd9f6c24027de177b2fb0eb6385542d62f4c21665a8d4cc7e1c118688e0836de"
+  "1.3.239.0":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.239.0.tar.gz"
+    sha256: "7aa7fb46e25e5ef0144d29c92122b631dc7c7c6804a6339f195b368ad53328e4"
+  "1.3.236.0":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.236.0.tar.gz"
+    sha256: "68f2cf70b1960f85e931ef56935e6ceda1beeb214f8fa319e6b95128b02b485a"
+  "1.3.231.1":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.231.1.tar.gz"
+    sha256: "ea40af0f499e7e97a86ee54410c5c78e7f7bac40f65ae09a1549773b6501bf4d"
+  "1.3.224.1":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.224.1.tar.gz"
+    sha256: "49c00e0119e3bc11e13c0c740e57c76b582b14f754f3779b85508c4d90d9df85"
+  "1.3.216.0":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.216.0.tar.gz"
+    sha256: "593d9b818d536490b70322a01b306ec165df5e7a70d770d05014fbd0b325fa15"
+  "1.3.211.0":
+    url: "https://github.com/KhronosGroup/Vulkan-ValidationLayers/archive/refs/tags/sdk-1.3.211.0.tar.gz"
+    sha256: "927c1cb98c81fe8a1a529cf2d977d701dcda49c495a19583dc00e178b6757203"
+patches:
+  "1.3.243.0":
+    - patch_file: "patches/1.3.243.0-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
+  "1.3.239.0":
+    - patch_file: "patches/1.3.239.0-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
+  "1.3.236.0":
+    - patch_file: "patches/1.3.236.0-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
+  "1.3.231.1":
+    - patch_file: "patches/1.3.231.1-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
+    - patch_file: "patches/1.3.231.1-0002-cmake-no-werror.patch"
+      patch_description: "Allow to disable Werror for old gcc/clang versions"
+      patch_type: "portability"
+  "1.3.224.1":
+    - patch_file: "patches/1.3.224.1-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
+  "1.3.216.0":
+    - patch_file: "patches/1.3.204.1-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
+  "1.3.211.0":
+    - patch_file: "patches/1.3.204.1-0001-fix-cmake.patch"
+      patch_description: "CMake: Adapt to conan"
+      patch_type: "conan"
