+++ recipes/vulkan-validationlayers/1.3.x/patches/1.3.204.1-0001-fix-cmake.patch
