+++ recipes/vulkan-validationlayers/1.3.x/patches/1.3.224.1-0001-fix-cmake.patch
