+++ recipes/vulkan-validationlayers/1.3.x/patches/1.3.231.1-0002-cmake-no-werror.patch
