+++ recipes/vulkan-validationlayers/1.3.x/test_package/test_package.cpp
@@ -1,0 +1,8 @@
+#include <vulkan/vk_format_utils.h>
+
+#include <iostream>
+
+int main() {
+  std::cout << "VK_FORMAT_D16_UNORM " << (FormatIsDepthOnly(VK_FORMAT_D16_UNORM) ? "is" : "is not") << " depth only" << std::endl;
+  return 0;
+}
