--- recipes/vulkan-validationlayers/4.x/test_package/CMakeLists.txt
@@ -1,7 +1,0 @@
-cmake_minimum_required(VERSION 3.15)
-project(test_package LANGUAGES CXX)
-
-find_package(vulkan-validationlayers REQUIRED CONFIG)
-
-add_executable(${PROJECT_NAME} test_package.cpp)
-target_link_libraries(${PROJECT_NAME} PRIVATE vulkan-validationlayers::vulkan-validationlayers)
