--- recipes/vulkan-validationlayers/4.x/test_package/test_package.cpp
@@ -1,6 +1,0 @@
-// Testing was performed manually. You expect to see validation layer messages in the output
-// where the VK_LAYER_KHRONOS_validation.dll is loaded if you create a vk instance
-// with the validation layers added
-int main() {
-   return 0;
-}
