+++ recipes/vulkan-validationlayers/all/test_package/test_package.cpp
@@ -1,8 +1,6 @@
-#include <vulkan/vk_format_utils.h>
-
-#include <iostream>
-
+// Testing was performed manually. You expect to see validation layer messages in the output
+// where the VK_LAYER_KHRONOS_validation.dll is loaded if you create a vk instance
+// with the validation layers added
 int main() {
-  std::cout << "VK_FORMAT_D16_UNORM " << (FormatIsDepthOnly(VK_FORMAT_D16_UNORM) ? "is" : "is not") << " depth only" << std::endl;
-  return 0;
+   return 0;
 }
