+++ recipes/vulkan-validationlayers/config.yml
@@ -1,17 +1,17 @@
 versions:
-  "1.4.341.0":
-    folder: 4.x
+  "1.4.350.0":
+    folder: all
   "1.3.243.0":
-    folder: all
+    folder: 1.3.x
   "1.3.239.0":
-    folder: all
+    folder: 1.3.x
   "1.3.236.0":
-    folder: all
+    folder: 1.3.x
   "1.3.231.1":
-    folder: all
+    folder: 1.3.x
   "1.3.224.1":
-    folder: all
+    folder: 1.3.x
   "1.3.216.0":
-    folder: all
+    folder: 1.3.x
   "1.3.211.0":
-    folder: all
+    folder: 1.3.x
